refactor(web): move both setup wizards onto the shared shell
The Slack and Telegram wizards each owned a Stepper, a sunken step panel,
step state, and a back-only step-select handler. The previous commit added
that as ChannelSetupWizard and useChannelSetupSteps; this converts both and
deletes the copies.

Each wizard now holds only what is actually its own: its steps, its channel's
handoff, and its credential shape.

Both also inherit the focus handling the shell brought with it, so a step
change no longer drops focus to <body> in either wizard. Verified in the
browser: activeElement is the step panel, labelled "Slack setup, step 2 of 4:
Open", where it was BODY before.

MutationStatus now comes from the shell. It was declared in the Slack wizard
and re-declared in the Telegram one, which left the type either duplicated or
imported across unrelated channels depending on which wart you picked.
`assistant-channels-list.tsx` and `channel-panel.tsx` follow it there.

Part of LUM-3077
